T2/llm_processor/requirements.txt
redis
requests
kafka-python==2.0.2